requirements.txt
ultralytics>=8.0.0
gtts>=2.3.0
Pillow>=10.0.0
streamlit>=1.28.0
torch>=2.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
requests>=2.31.0

